Fixing issue which does not compute effect of rotor inertia

+++ src/RBDyn/RBDyn/MultiBody.h
@@ -216,12 +216,12 @@ public:
 
   void setJointRotorInertia(int num, double Ir)
   {
-    joints_[num].setRotorInertia(Ir);
+    joints_[static_cast<std::size_t>(num)].setRotorInertia(Ir);
   }
 
   void setJointGearRatio(int num, double gr)
   {
-    joints_[num].setGearRatio(gr);
+    joints_[static_cast<std::size_t>(num)].setGearRatio(gr);
   }
 
   // safe accessors version for python binding
